separate into utils
